server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
networkx
matplotlib
openai
